requirements.txt
flask
flask-cors
bcrypt
psycopg2-binary
requests
werkzeug
